python_audio/requirements.txt
openai-whisper
speechbrain
torch
torchaudio
fastapi
uvicorn[standard]
python-multipart
numpy
soundfile
